third_party/blink/web_tests/http/tests/xmlhttprequest/duplicate-revalidation-reload-expected.txt

Make two XHRs for the resource which is already cached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result1 is "PASS"
PASS result2 is "PASS"
PASS DONE
PASS successfullyParsed is true

TEST COMPLETE
